refactor vote referendum modal
